fix ticket 1028

+++ src/Module/Colony/Lib/BuildPlanDeleter.php
@@ -4,11 +4,15 @@ declare(strict_types=1);
 
 namespace Stu\Module\Colony\Lib;
 
+use InvalidArgumentException;
+use Stu\Module\PlayerSetting\Lib\UserConstants;
 use Stu\Orm\Entity\SpacecraftBuildplan;
 use Stu\Orm\Repository\BuildplanModuleRepositoryInterface;
 use Stu\Orm\Repository\ColonyShipQueueRepositoryInterface;
+use Stu\Orm\Repository\DealsRepositoryInterface;
 use Stu\Orm\Repository\ShipyardShipQueueRepositoryInterface;
 use Stu\Orm\Repository\SpacecraftBuildplanRepositoryInterface;
+use Stu\Orm\Repository\UserRepositoryInterface;
 
 /**
  * Provides service methods for ship buildplan deletion
@@ -19,14 +23,33 @@ final class BuildPlanDeleter implements BuildPlanDeleterInterface
         private SpacecraftBuildplanRepositoryInterface $spacecraftBuildplanRepository,
         private BuildplanModuleRepositoryInterface $buildplanModuleRepository,
         private ColonyShipQueueRepositoryInterface $colonyShipQueueRepository,
-        private ShipyardShipQueueRepositoryInterface $shipyardShipQueueRepository
+        private ShipyardShipQueueRepositoryInterface $shipyardShipQueueRepository,
+        private DealsRepositoryInterface $dealsRepository,
+        private UserRepositoryInterface $userRepository
     ) {}
 
     #[\Override]
     public function delete(SpacecraftBuildplan $spacecraftBuildplan): void
     {
-        $this->buildplanModuleRepository->truncateByBuildplan($spacecraftBuildplan->getId());
+        $buildplanId = $spacecraftBuildplan->getId();
+        if ($this->dealsRepository->hasBuildplan($buildplanId)) {
+            $this->transferToForeignBuildplansUser($spacecraftBuildplan);
+            return;
+        }
+
+        $this->buildplanModuleRepository->truncateByBuildplan($buildplanId);
         $this->spacecraftBuildplanRepository->delete($spacecraftBuildplan);
+    }
+
+    private function transferToForeignBuildplansUser(SpacecraftBuildplan $spacecraftBuildplan): void
+    {
+        $user = $this->userRepository->find(UserConstants::USER_FOREIGN_BUILDPLANS);
+        if ($user === null) {
+            throw new InvalidArgumentException(sprintf('user with id %d does not exist', UserConstants::USER_FOREIGN_BUILDPLANS));
+        }
+
+        $spacecraftBuildplan->setUser($user);
+        $this->spacecraftBuildplanRepository->save($spacecraftBuildplan);
     }
 
     #[\Override]
